backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
torch>=2.2.0
torchvision>=0.17.0
onnx>=1.16.0
onnxruntime>=1.18.0
numpy>=1.26.0
Pillow>=10.0.0
transformers>=4.40.0
websockets>=12.0
python-multipart>=0.0.9
